feat: add dummy methods for "BaseClient" in efb-qq-slave
We need to first set up all the abstract methods when deriving the
"BaseClient" defined in efb-qq-slave to initialize the class.

In order to run the ehForwarderBot at now, we have to define the "poll"
and "stop_poll" method. However, we use the aiocqhttp as the dependency,
it is an asynchronous library, but EFB is always synchronous.

EFB will create a thread for each slave instance to call "poll" method
to star the slave. However, we need to create a new thread to isolate
the event loop. Otherwise, there may be some unexpected behaviors.

EBF will call "stop_polling" method to stop the slave instance. However,
we cannot simply use "self.event_loop.stop()" here, because when calling
"event_loop.stop" outside the thread where the event loop is running, it
may cause deadlock. So we need to use "call_soon_threadsafe" to execute
"event_loop.stop" action to avoid deadlock.