napcat: set up exceptions
Create a new exception "NapCatException" as the base class for all the
exceptions. This is important, because we can use polymorphism to make
the code more clean.

We first create "NapCatAPIFailureExtension", it is used to encapsulate
the aiocqhttp non-network failure. And we include "_status_code" and
"_ret_code" in the exception to avoid repeated code when initializing
the exception.

Then, we add others exceptions for better semantic representation.