napcat: set up the basic bot operations
It's not a good idea to let "NapCatBot" class to implement too many
functionalities. We need to use composition method to make the code more
clean.

Because we are a plugin of the efb-qq-slave. The configuration will be
passed from the "efb-qq-slave". At current, we support the following
configs:

```yaml
Client: NapCat
NapCat:
  api_root: http://192.168.100.132/6700
  access_token: ""
  api_timeout: 60
```

And we use "_call_action_wrapper" to wrap the aiocqhttp "call_action"
method, it will handle the exceptions raised by the aiocqhttp and raise
the following exceptions when failing:

1. NapCatDisconnectedException: The NapCat client is disconnected.
2. NapCatAPIFailureException: The NapCat HTTP API fails.

Then we use "call_action" to wrap the "_call_action_wrapper", it will
first check the "self._logged_in" and "self._connected" field. It not,
it should warn the user with limited times. Currently, the code is NOT
implemented.

Then, we provide a function "check_status_periodically" to set up the
"self._logged_in" and "self._connected" field. If anything is bad, we
need to warn the user. However, currently we do NOT implement the logic
how to warn the user.

At last, add unit tests.