napcat: add friend manager
Add "friend_manager" class to support friend operations. It will call
the "$(api_root)/get_friend_list" to get all the friends which is stored
in the "friend_manager._friend_list". And if the "remark" field of the
"Friend" is empty, we will use the "nickname" instead.

We will also create a mapping from the "user_id" to the "Friend"
instance. It's important here for us to quickly find the "Friend" by the
"user_id". This is because "user_id" is unique, we will use this unique
identifier in the EBF chat later.

Last, provide a function to get a remark of a friend, if the friend is
not in the "friend_manager._friend_list", we will fetch the newest
friend lists from the http server. If we cannot find still, I guess you
do not have a such friend.