napcat: avoid assign the whole response to friend_list
In the original code, we simply assign the response to the "_friend_list".
However, there are many unused fields in the response and we should not
restore these contents in the "_friend_list" which may cost a lot of
memory usage.

Rename "_populate_uid_to_friend_mapping" to "_update_friend_list_callback"
to allow more meaningful meanings.

At last, rename a bad test name and add a new test to test the situation
where the "remark" field is empty we should use "nickname" field
instead.